feature: Option의  min,max를 OptionGroup으로 이동
