+++ src/main/java/com/example/freshcart/product/domain/OptionGroupRepository.java
@@ -1,10 +1,9 @@
 package com.example.freshcart.product.domain;
 
-import java.util.List;
 
 public interface OptionGroupRepository {
 
   OptionGroup save(OptionGroup optionGroup);
 
-  List<OptionGroup> findByProductId(Long id);
+  OptionGroup findById(Long productOptionGroupId);
 }
