+++ src/main/java/com/example/freshcart/product/infrastructure/OptionGroupMapper.java
@@ -1,7 +1,6 @@
 package com.example.freshcart.product.infrastructure;
 
 import com.example.freshcart.product.domain.OptionGroup;
-import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
@@ -9,5 +8,5 @@ public interface OptionGroupMapper {
 
   void insert(OptionGroup optionGroup);
 
-  List<OptionGroup> findByProductId(Long id);
+  OptionGroup findById(Long id);
 }
