+++ src/main/resources/mapper/option-group-mapper.xml
@@ -3,16 +3,24 @@
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.example.freshcart.product.infrastructure.OptionGroupMapper">
   <insert id="insert" keyProperty="id" useGeneratedKeys="true">
-    insert into option_group (name, is_required, is_count_required, product_id)
-    values (#{optionGroupName}, #{isRequired}, #{isCountRequired}, #{productId});
+    insert into option_group (name, is_required, exclusive, minimum_order, maximum_order, product_id)
+    values (#{optionGroupName}, #{isRequired}, #{exclusive}, #{minimumOrder}, #{maximumOrder}, #{productId});
   </insert>
 
   <resultMap id="optionGroupResultMap" type="com.example.freshcart.product.domain.OptionGroup">
+    <result column="name" property="optionGroupName"/>
     <result column="is_required" property="isRequired"/>
-    <result column="is_count_required" property="isCountRequired"/>
+    <result column="minimum_order" property="minimumOrder"/>
+    <result column="maximum_order" property="maximumOrder"/>
     <result column="product_id" property="productId"/>
     <result column="created_at" property="createdAt"/>
     <result column="updated_at" property="updatedAt"/>
   </resultMap>
+
+  <select id="findById" resultMap="optionGroupResultMap">
+    select id, name, is_required, exclusive, minimum_order, maximum_order, product_id, created_at, updated_at
+    from option_group
+    where id = #{productOptionGroupId};
+  </select>
 
 </mapper>
